arctan as a power series
- Definition of arctan as a power series throught reduction of
the argument and oddity equations.

- Neat (and short!) proof that arctan is derivable on D(0,1)
with value / (1 + x ^ 2) and extension to the whole R+ \ 1.

- Proof that arctan and atan are equal on D(0,1). To get the
proof that atan 1 = arctan 1 (necessary to prove sin PI/2 = 1
but also to unlock atan == arctan on the whole real line), we
need to prove that arctan is continuous in 1.

There's a mediocre proof in my L3 internship but we can probably
find something more general / cleaner. Is it true that if the
power series converges on a point on the edge of the disk then
the function "sum" is continuous at this point (R case only of
course!). i.e. is this goal solvable?

An  : Rseq
r   : R
rho : finite_cv_radius An r
l   : R
Hl  : Rpser An r l
==========================================
continuity_pt_in (interval 0 r) (sum_r An r rho) r



git-svn-id: svn+ssh://svn.code.sf.net/p/coqtail/code@400 7194d10d-a3ed-4628-aea2-7b4a0dfde000